some fixes

+++ bot.py
@@ -202,17 +202,17 @@ async def on_message(message):
                 await bot.get_channel(516742583808950272).send("<@277981712989028353>", embed=em)
     
     # Portal
-    if message.channel.id == 566030691360440356: # DBBBH Side
+    if message.channel.id == 566030691360440356 and not message.author.discriminator == "0000": # DBBBH Side
         img = message.author.avatar_url_as(format="png", size=1024)
         byte = await (await bot.session.get(img)).read()
         webhook = await bot.get_channel(566015775261982866).create_webhook(name=message.author.display_name, avatar=byte)
-        await webhook.send(message)
+        await webhook.send(message.content.replace("@", "@\u200b"))
         await webhook.delete()
-    elif message.channel.id == 566015775261982866: # Pixel side
+    elif message.channel.id == 566015775261982866 and not message.author.discriminator == "0000": # Pixel side
         img = message.author.avatar_url_as(format="png", size=1024)
         byte = await (await bot.session.get(img)).read()
         webhook = await bot.get_channel(566030691360440356).create_webhook(name=message.author.display_name, avatar=byte)
-        await webhook.send(message)
+        await webhook.send(message.content.replace("@", "@\u200b"))
         await webhook.delete()
         
     # Blacklist
